Moved ReceiptService.go to service.go and cleaned up some code
