+++ internal/app/commands/payment/receipt/command_default.go
@@ -13,6 +13,6 @@ func (c *RCommander) Default(inputMessage *tgbotapi.Message) {
 
 	_, err := c.bot.Send(msg)
 	if err != nil {
-		log.Printf("PaymentReceiptCommander.Help: error sending reply message to chat - %v", err)
+		log.Printf("RCommander.Help: error sending reply message to chat - %v", err)
 	}
 }
